requirements.txt
ccxt
numpy
pandas
lightgbm
scikit-learn
joblib
tensorflow
schedule
python-dotenv 